@@ -28,6 +28,8 @@
     "@ngxs/devtools-plugin": "^3.0.1",
     "@ngxs/store": "^3.0.1",
     "@types/nedb": "^1.8.5",
+    "@types/pouchdb-core": "^6.4.0",
+    "@types/pouchdb-find": "^6.3.2",
     "a-color-picker": "^1.0.5",
     "babel-polyfill": "^6.26.0",
     "bootstrap": "^4.1.1",
@@ -40,8 +42,11 @@
     "ngx-color": "^1.3.1",
     "ngx-smart-modal": "^6.0.0",
     "ngx-toastr": "^8.5.1",
+    "pouchdb": "^6.4.3",
     "pouchdb-adapter-idb": "^6.4.3",
+    "pouchdb-find": "^6.4.3",
     "pretty-checkbox": "^3.0.3",
+    "relational-pouch": "^3.0.1",
     "rxdb": "^7.6.0",
     "rxjs": "^6.0.0",
     "zone.js": "^0.8.26"
@@ -57,6 +62,7 @@
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
+    "@types/pouchdb": "^6.3.2",
     "codelyzer": "~4.2.1",
     "husky": "^0.14.3",
     "jasmine-core": "~2.99.1",
@@ -74,6 +80,6 @@
     "stylelint-scss": "^3.1.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "typescript": "^2.9.1-insiders.20180525"
   }
 }
