@@ -11,53 +11,10 @@ import { RxCollection, RxDatabase } from 'rxdb';
   template: '<router-outlet></router-outlet>'
 })
 export class InkApp implements OnInit {
-  constructor(private store: Store, private db: DBService, private zone: NgZone) {
+  constructor(private store: Store, private db: DBService) {
     // console.log('cons', this.db.conn);
   }
   ngOnInit() {
-    try {
-      this.getDb();
-    } catch (e) {
-      console.log('Unexpected error occurred', e);
-    }
     this.store.dispatch(new ChangeView(InkAppView.ROUND));
-  }
-
-  async getDb() {
-    this.zone.run(() => {});
-    const db = await this.db.localDb.getDatabase();
-    // const d = await db.board.dump(false);
-    console.log('db');
-    // const firstDoc = await db.board.findOne().exec();
-    // console.log('name', firstDoc);
-    // const f: string = firstDoc.name;
-
-    // this.db.localDb
-    //   .getDatabase()
-    //   .then((a: InkDb) => {
-    //     a.board.$.subscribe(f => console.log(f));
-    //     // const foo = a.board.newDocument({ id: 'xyz', name: 'super appu' });
-    //     // foo.save();
-    //     return a;
-    //   })
-    //   .then(db => {
-    //     db.board
-    //       .find()
-    //       .exec()
-    //       .catch();
-    //     console.log();
-    //     // const fgt = db.board.find().$.subscribe(
-    //     //   n => {
-    //     //     console.log('now', n);
-    //     //   },
-    //     //   e => {
-    //     //     console.log('de', e);
-    //     //   }
-    //     // );
-    //     return db;
-    //   })
-    //   .catch(e => {
-    //     console.log('Unexpected error occurred', e);
-    //   });
   }
 }
