@@ -1,4 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { BoardService } from '../../services/board.service';
+import { Store } from '@ngxs/store';
+import { LoadBoard } from '../../store/actions/board.actions';
+import { filter, map } from 'rxjs/operators';
+import { InkBoard, InkBoardMeta } from '../../models';
 
 @Component({
   selector: 'inkapp-home-page',
@@ -6,7 +11,15 @@ import { Component, OnInit } from '@angular/core';
   styles: []
 })
 export class HomePage implements OnInit {
-  constructor() {}
+  boards: InkBoardMeta[];
+  constructor(private _boardService: BoardService, private _store: Store) {}
 
-  ngOnInit() {}
+  async ngOnInit() {
+    this._store.select(s => s.board).subscribe(b => {
+      this.boards = b;
+    });
+    const boards = await this._boardService.getBoards();
+    boards.map(b => ({ _id: b._id, name: b.name, description: b.description, createdAt: b.createdAt }));
+    this._store.dispatch(new LoadBoard(boards));
+  }
 }
